requirements.txt
beautifulsoup4==4.11.1
discord.py==2.1.0
pandas==1.5.1
requests==2.28.0
selenium==4.6.0
webdriver_manager==3.7.0
